% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mean_n_at.R
\name{add_mean_n_at}
\alias{add_mean_n_at}
\title{Add mean number of at signs}
\usage{
add_mean_n_at(output, hierarchies, include_HiTaS_Class = TRUE)
}
\arguments{
\item{output}{Data frame. Output from a suppression function.}

\item{hierarchies}{List of data frames as dimList in sdcTable}

\item{include_HiTaS_Class}{Logical; whether to include the HiTaS classification. Default is TRUE.}
}
\description{
Add mean number of at signs
}
\examples{

d3 <- SSBtools::SSBtoolsData("d3")
d3$value <- (d3$freq + pi)^2
hi <- SSBtools::FindDimLists(d3[c(1:2, 4:6)])
a <- GaussSuppression::SuppressDominantCells(d3, 
                           dominanceVar = "value", 
                           hierarchies = hi, 
                           pPercent = pi, 
                           singletonMethod = "numttTtT")

b <- add_mean_n_at(a, hi)
head(b)
tail(b)

# if pipe operator is preferred (|> can be replaced by \%>\%)
b <- a |> add_mean_n_at(hi)

}
