% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_response.R
\name{add_response}
\alias{add_response}
\title{Add a randomly generated response variable to data}
\usage{
add_response(df, shape = 1.1, response = "response", rnd_seed = 123)
}
\arguments{
\item{df}{data frame}

\item{shape}{shape parameter}

\item{response}{name of response variable}

\item{rnd_seed}{If non-NULL, a random generator seed to be used locally 
within the function without affecting the random value stream in R.}
}
\description{
Pareto Type II (Lomax) whole nonzero random numbers are added.
}
\examples{
df <- SSBtools::SSBtoolsData("paris2025_micro")

add_response(df)

# if pipe operator is preferred (|> can be replaced by \%>\%)
df |> add_response()
df |> add_response() |> add_response(shape = 2, response = "r2")
df |> add_response() |> add_response(rnd_seed = 44, response = "r44")
}
