% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_primary.R
\name{check_primary}
\alias{check_primary}
\title{Run and Check Primary Suppression}
\usage{
check_primary(
  df_microdata,
  hierarchies,
  pvalue = 5,
  print_info = TRUE,
  do_check_sdcTable = FALSE
)
}
\arguments{
\item{df_microdata}{A data frame of generated microdata.}

\item{hierarchies}{A list of hierarchies corresponding to the dimensions in the microdata.}

\item{pvalue}{Numeric; threshold for identifying unsafe cells based on the p-percent rule. Default is 5.}

\item{print_info}{Logical; if TRUE, prints summary statistics of the primary suppression. Default is TRUE.}

\item{do_check_sdcTable}{Logical; if TRUE, cross-checks results against `sdcTable::primarySuppression`. Default is FALSE.}
}
\value{
A list containing the processed SDC list object with primary suppression results, and optionally sdcTable verification metrics.
}
\description{
Runs primary suppression on microdata and compares/adds results from info_output.
}
