% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_primary.R
\name{check_sdcTable}
\alias{check_sdcTable}
\title{Compare Primary Suppression with sdcTable}
\usage{
check_sdcTable(df_merged, df_microdata, hierarchies, pvalue = 5)
}
\arguments{
\item{df_merged}{A data frame containing the precomputed merged/suppressed data.}

\item{df_microdata}{A data frame of generated microdata.}

\item{hierarchies}{A list of hierarchies corresponding to the dimensions in the microdata.}

\item{pvalue}{Numeric; threshold for identifying unsafe cells. Default is 5.}
}
\value{
A list with the output data frame from sdcTable, a logical vector of its primary suppressions, and an equality check result.
}
\description{
Runs primary suppression using `sdcTable::primarySuppression` to verify that the results match the package's precomputed primary suppressions.
}
