% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_microdata.R
\name{create_microdata}
\alias{create_microdata}
\title{Generate Mock Microdata}
\usage{
create_microdata(
  hiers,
  n_ids = 1000,
  n_unique = n_ids/10,
  shape = 1.1,
  prob_dim = 1/seq_len(1e+05),
  prob_freq = 1/seq_len(1e+05),
  rnd_seed = 123,
  sample_seed = 123
)
}
\arguments{
\item{hiers}{A list of hierarchy structures.}

\item{n_ids}{Total number of observations (rows) to generate. Default is 1000.}

\item{n_unique}{Number of unique dimension combinations to generate. Default is n_ids/10.}

\item{shape}{Shape parameter for the Pareto Type II (Lomax) distribution of cell responses. Default is 1.1.}

\item{prob_dim}{Probability vector for dimension levels.}

\item{prob_freq}{Probability vector for generating cell frequencies.}

\item{rnd_seed}{Random seed for cell response generation. Default is 123.}

\item{sample_seed}{Random seed for sampling the combinations. Default is 123.}
}
\value{
A data frame containing the generated mock microdata.
}
\description{
Generates mock microdata representing individual/transactional records with hierarchical dimensions and responses.
}
