% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_microdata.R
\name{info_microdata}
\alias{info_microdata}
\title{Get Information About Microdata}
\usage{
info_microdata(df_microdata, hierarchies, return_frame = FALSE, pvalue = 5)
}
\arguments{
\item{df_microdata}{A data frame of generated microdata.}

\item{hierarchies}{A list of hierarchies corresponding to the dimensions in the microdata.}

\item{return_frame}{Logical; if TRUE, returns the underlying frequencies and p-percent values. Default is FALSE.}

\item{pvalue}{The threshold for identifying unsafe cells based on the p-percent rule. Default is 5.}
}
\value{
Prints summary statistics. If `return_frame` is TRUE, returns a data frame with frequencies and p-percent values.
}
\description{
Computes and prints key metrics of the generated microdata, such as empty inner cells, singleton cells, and unsafe cells.
}
