% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_response.R
\name{rpareto2}
\alias{rpareto2}
\title{Pareto Type II (Lomax) whole nonzero random numbers}
\usage{
rpareto2(n, shape = 1.1, max_response = 10^9 - 1, cuttail = 1e-07)
}
\arguments{
\item{n}{number of observations}

\item{shape}{shape parameter}

\item{max_response}{preset value on largest observation. This automatically determines the scale.}

\item{cuttail}{Values above the (1 - cuttail) quantile are truncated.
This prevents extremes from giving rise to a large proportion of ones.}
}
\description{
Modified version of VGAM::rparetoII
}
\examples{
rpareto2(5)
rpareto2(5, shape = 2)
}
